Added Religious Affiliation History theme
Currently I have no idea how to figure out which religion is which, so
on the map they're just labeled as "Religion1" - "Religion4".

Another thing, there's a NaN that gets logged in the console every time
you click on the slider. It happens on the History theme so I don't
think it's something I accidently broke. It looks like the error occurs
when it's calculating the offset to scroll to the next story.