Changes to Railroads
Changes to Query have been put in but due to the multiple child layers it has been difficult to control which layer is viewing and set the right visible layer. The same issue I believe that is partly wrong with the other themes. Lot of work that still needs to be done on railroads.