Changes to Railroads
I have created a new and improved map in order to improve design. Rails are finally showing up on the map but still running into Query issues. Still requries a lot of work. Also the only layer visible is the rails, I need to put the counties back in for better reprsentation.