Changes to Railroads
Railroads.js now has a working display of railroads when the user changes the year on the slider. After creating my own test railroad file I have found the source of the issue for the orignial railroad project. Hiding the orginal features and creating a new line symbol that could in fact "draw" the features selected was used. More improvments are needed to be made but the biggest issue has been figured out.