Railroads Narritve has been added
The railroads narrative has finally been added to the database. Currently working to link the narrative to the railroads.js file when called upon. Currently when clicked on the narrative will only show up with no map loading due to it trying to use the maps.js file other than the railroads.js file. Currently looking into the server.js file to see if changes can be made to fix this issue.