County Churches Theme now availiable
The county churches theme is still yet in the works but here is my progess so far: Map can and is successfully queried for what ever year the user chooses, map displays all churches under that given year, and updated style topo map (just as in railroads) is used. More work needs to be done to make this map more informative but sturcture is now set in stone.