County Churches and Railroads are updated
Railroad counties layer have now been added to show the changes in county bounderies as well as the changes seen with the rails within those counties. County Churches now successfully renderers the images found within the MapServer in order to accuratly display data on the map. Just the same as in railroads, county bounderies will also change depending on what year is choosen. County Churches is ready to be put live on the narrative is created.